requirements.txt
fastapi
uvicorn
aiohttp
asyncpg
sqlalchemy
pydantic-settings
openai
numpy
jinja2
beautifulsoup4
python-dotenv
Pillow
lxml
crawl4ai
pyyaml
